# static/imdbapi.swagger.yaml
swagger: "2.0"
info:
  title: IMDxAPI
  description: IMDx API for accessing movie and TV show data
  version: 2.7.12
  contact:
    name: IMDxAPI Source Code
    url: https://github.com/badman99dev/IMDxAPI
...
consumes:
  - application/json
produces:
  - application/json
paths:
  /titles:
    get:
      summary: List titles
      description: Retrieve a list of titles with optional filters.
      operationId: IMDbAPIService_ListTitles
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiListTitlesResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: types
          description: |-
            Optional. The type of titles to filter by.
            If not specified, all types are returned.

             - MOVIE: Represents a movie title.
             - TV_SERIES: Represents a TV series title.
             - TV_MINI_SERIES: Represents a TV mini-series title.
             - TV_SPECIAL: Represents a TV special title.
             - TV_MOVIE: Represents a TV movie title.
             - SHORT: Represents a short title.
             - VIDEO: Represents a video title.
             - VIDEO_GAME: Represents a video game title.
          in: query
          required: false
          type: array
          items:
            type: string
            enum:
              - MOVIE
              - TV_SERIES
              - TV_MINI_SERIES
              - TV_SPECIAL
              - TV_MOVIE
              - SHORT
              - VIDEO
              - VIDEO_GAME
          collectionFormat: multi
        - name: genres
          description: |-
            Optional. The genres to filter titles by.
            If not specified, titles from all genres are returned.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: countryCodes
          description: |-
            Optional. The ISO 3166-1 alpha-2 country codes to filter titles by.
            If not specified, titles from all countries are returned.
            Example: "US" for United States, "GB" for United Kingdom.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: languageCodes
          description: |-
            Optional. The ISO 639-1 or ISO 639-2 language codes to filter titles by.
            If not specified, titles in all languages are returned.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: nameIds
          description: Optional. The IDs of names to filter titles by.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: interestIds
          description: |-
            Optional. The IDs of interests to filter titles by.
            If not specified, titles associated with all interests are returned.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: startYear
          description: Optional. The start year for filtering titles.
          in: query
          required: false
          type: integer
          format: int32
        - name: endYear
          description: Optional. The end year for filtering titles.
          in: query
          required: false
          type: integer
          format: int32
        - name: minVoteCount
          description: |-
            Optional. The minimum number of votes a title must have to be included.
            If not specified, titles with any number of votes are included.
            The value must be between 0 and 1,000,000,000.
            Default is 0.
          in: query
          required: false
          type: integer
          format: int32
        - name: maxVoteCount
          description: |-
            Optional. The maximum number of votes a title can have to be included.
            If not specified, titles with any number of votes are included.
            The value must be between 0 and 1,000,000,000.
          in: query
          required: false
          type: integer
          format: int32
        - name: minAggregateRating
          description: |-
            Optional. The minimum rating a title must have to be included.
            If not specified, titles with any rating are included.
            The value must be between 0.0 and 10.0.
          in: query
          required: false
          type: number
          format: float
        - name: maxAggregateRating
          description: |-
            Optional. The maximum rating a title can have to be included.
            If not specified, titles with any rating are included.
            The value must be between 0.0 and 10.0.
          in: query
          required: false
          type: number
          format: float
        - name: sortBy
          description: |-
            Optional. The sorting order for the titles.
            If not specified, titles are sorted by popularity.

             - SORT_BY_POPULARITY: Sort by popularity.
            This is used to rank titles based on their popularity, which can be influenced by various factors
            such as viewership, ratings, and cultural impact.
             - SORT_BY_RELEASE_DATE: Sort by release date.
            This is used to rank titles based on their release dates, with newer titles typically appearing
            before older ones.
             - SORT_BY_USER_RATING: Sort by user rating.
            This is used to rank titles based on the average user rating, which reflects the overall audience reception.
             - SORT_BY_USER_RATING_COUNT: Sort by user rating count.
            This is used to rank titles based on the number of user ratings they have received,
            which can indicate the level of engagement or popularity among viewers.
             - SORT_BY_YEAR: Sort by year.
            This is used to rank titles based on their release year, with newer titles typically appearing before older ones.
          in: query
          required: false
          type: string
          enum:
            - SORT_BY_POPULARITY
            - SORT_BY_RELEASE_DATE
            - SORT_BY_USER_RATING
            - SORT_BY_USER_RATING_COUNT
            - SORT_BY_YEAR
        - name: sortOrder
          description: |-
            Optional. The sorting order for the titles.
            If not specified, titles are sorted in ascending order.

             - ASC: Sort in ascending order.
             - DESC: Sort in descending order.
          in: query
          required: false
          type: string
          enum:
            - ASC
            - DESC
        - name: pageToken
          description: Optional. Token for pagination, if applicable.
          in: query
          required: false
          type: string
      tags:
        - Title
  /titles/{titleId}:
    get:
      summary: Get title by ID
      description: Retrieve a title's details using its IMDb ID.
      operationId: IMDbAPIService_GetTitle
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiTitle'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleId
          description: The IMDb title ID in the format 'tt1234567'.
          in: path
          required: true
          type: string
      tags:
        - Title
  /titles:batchGet:
    get:
      summary: Batch get titles by IDs
      description: Retrieve details of multiple titles using their IMDb IDs.
      operationId: IMDbAPIService_BatchGetTitles
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiBatchGetTitlesResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleIds
          description: List of IMDb title IDs. Maximum 5 IDs.
          in: query
          required: true
          type: array
          items:
            type: string
          collectionFormat: multi
      tags:
        - Title
  /titles/{titleId}/credits:
    get:
      summary: List credits for a title
      description: Retrieve the credits associated with a specific title.
      operationId: IMDbAPIService_ListTitleCredits
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiListTitleCreditsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleId
          description: Required. IMDb title ID in the format "tt1234567".
          in: path
          required: true
          type: string
        - name: categories
          description: Optional. The categories of credits to filter by.
          in: query
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: pageSize
          description: |-
            Optional. The maximum number of credits to return per page.
            If not specified, a default value will be used.

            The value must be between 1 and 50. Default is 20.
          in: query
          required: false
          type: integer
          format: int32
        - name: pageToken
          description: Optional. Token for pagination, if applicable.
          in: query
          required: false
          type: string
      tags:
        - Title
  /titles/{titleId}/releaseDates:
    get:
      summary: List release dates for a title
      description: Retrieve the release dates associated with a specific title.
      operationId: IMDbAPIService_ListTitleReleaseDates
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiListTitleReleaseDatesResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleId
          description: Required. IMDb title ID in the format "tt1234567".
          in: path
          required: true
          type: string
        - name: pageSize
          description: |-
            Optional. The maximum number of credits to return per page.
            If not specified, a default value will be used.

            The value must be between 1 and 50. Default is 20.
          in: query
          required: false
          type: integer
          format: int32
        - name: pageToken
          description: Optional. Token for pagination, if applicable.
          in: query
          required: false
          type: string
      tags:
        - Title
  /titles/{titleId}/akas:
    get:
      summary: List AKAs for a title
      description: Retrieve the alternative titles (AKAs) associated with a specific title.
      operationId: IMDbAPIService_ListTitleAKAs
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiListTitleAKAsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleId
          description: Required. IMDb title ID in the format "tt1234567".
          in: path
          required: true
          type: string
      tags:
        - Title
  /titles/{titleId}/seasons:
    get:
      summary: List seasons for a title
      description: Retrieve the seasons associated with a specific title.
      operationId: IMDbAPIService_ListTitleSeasons
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/imdbapiListTitleSeasonsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: titleId
          description: Required. IMDb title ID in the format "tt1234567".
          in: path
          required: true
          type: string
      tags:
        - Title
...
definitions:
...
  imdbapiBatchGetTitlesResponse:
    type: object
    properties:
      titles:
        type: array
        items:
          type: object
          $ref: '#/definitions/imdbapiTitle'
        description: List of titles retrieved by their IMDb IDs.
...
  imdbapiListTitleCreditsResponse:
    type: object
    properties:
      credits:
        type: array
        items:
          type: object
          $ref: '#/definitions/imdbapiCredit'
        description: List of credits associated with the title.
      totalCount:
        type: integer
        format: int32
        description: Total count of credits associated with the title.
      nextPageToken:
        type: string
        description: Token for the next page of results, if applicable.
    description: Response message for listing credits associated with a title.
...
  imdbapiListTitlesResponse:
    type: object
    properties:
      titles:
        type: array
        items:
          type: object
          $ref: '#/definitions/imdbapiTitle'
        description: List of titles matching the request criteria.
      totalCount:
        type: integer
        format: int32
        description: Total number of titles matching the request criteria.
      nextPageToken:
        type: string
        description: Token for the next page of results, if applicable.
    description: Response message for listing titles.
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: { }
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'